Split Setting into interface and implementation
