+++ engine-tests/src/test/java/org/terasology/config/flexible/FlexibleConfigTest.java
@@ -41,10 +41,10 @@ public class FlexibleConfigTest {
             SimpleUri id1 = new SimpleUri("engine-tests:TestSetting1");
             SimpleUri id2 = new SimpleUri("engine-tests:TestSetting2");
 
-            config.add(new Setting<>(id1, 50,
+            config.add(new SettingImpl<>(id1, 50,
                     new RangedNumberValidator<>(0, 100, false, false)));
 
-            config.add(new Setting<>(id2, 25d,
+            config.add(new SettingImpl<>(id2, 25d,
                     new RangedNumberValidator<>(0d, 100d, false, false)));
 
             Setting<Integer> setting1 = config.get(id1);
@@ -67,7 +67,7 @@ public class FlexibleConfigTest {
         public void testHas() throws Exception {
             SimpleUri id = new SimpleUri("engine-tests:TestSetting");
 
-            config.add(new Setting<>(id, 50,
+            config.add(new SettingImpl<>(id, 50,
                     new RangedNumberValidator<>(0, 100, false, false)));
 
             assertTrue(config.has(id));
@@ -77,7 +77,7 @@ public class FlexibleConfigTest {
         public void testNotHas() throws Exception {
             SimpleUri id = new SimpleUri("engine-tests:TestSetting");
 
-            config.add(new Setting<>(id, 50,
+            config.add(new SettingImpl<>(id, 50,
                     new RangedNumberValidator<>(0, 100, false, false)));
 
             assertFalse(config.has(KEY_NON_EXISTENT));
@@ -96,7 +96,7 @@ public class FlexibleConfigTest {
         public void testAdd() throws Exception {
             SimpleUri id = new SimpleUri("engine-tests:TestSetting");
 
-            assertNotNull(config.add(new Setting<>(id, 50,
+            assertNotNull(config.add(new SettingImpl<>(id, 50,
                     new RangedNumberValidator<>(0, 100, false, false))));
         }
 
@@ -104,10 +104,10 @@ public class FlexibleConfigTest {
         public void testAddExisting() throws Exception {
             SimpleUri id = new SimpleUri("engine-tests:TestSetting");
 
-            assertTrue(config.add(new Setting<>(id, 50,
+            assertTrue(config.add(new SettingImpl<>(id, 50,
                     new RangedNumberValidator<>(0, 100, false, false))));
 
-            assertFalse(config.add(new Setting<>(id, 5d,
+            assertFalse(config.add(new SettingImpl<>(id, 5d,
                     new RangedNumberValidator<>(1d, 100d, false, false))));
         }
     }
@@ -125,10 +125,10 @@ public class FlexibleConfigTest {
             SimpleUri id1 = new SimpleUri("engine-tests:TestSetting1");
             SimpleUri id2 = new SimpleUri("engine-tests:TestSetting2");
 
-            config.add(new Setting<>(id1, 50,
+            config.add(new SettingImpl<>(id1, 50,
                     new RangedNumberValidator<>(0, 100, false, false)));
 
-            config.add(new Setting<>(id2, 25d,
+            config.add(new SettingImpl<>(id2, 25d,
                     new RangedNumberValidator<>(0d, 100d, false, false)));
 
             assertTrue(config.remove(id1));
@@ -139,7 +139,7 @@ public class FlexibleConfigTest {
         public void testNonexistentRemove() throws Exception {
             SimpleUri id = new SimpleUri("engine-tests:TestSetting");
 
-            config.add(new Setting<>(id, 50,
+            config.add(new SettingImpl<>(id, 50,
                     new RangedNumberValidator<>(0, 100, false, false)));
 
             assertFalse(config.remove(KEY_NON_EXISTENT));
@@ -149,7 +149,7 @@ public class FlexibleConfigTest {
         public void testSubscribedRemove() throws Exception {
             SimpleUri id = new SimpleUri("engine-tests:TestSetting");
 
-            config.add(new Setting<>(id, 50,
+            config.add(new SettingImpl<>(id, 50,
                     new RangedNumberValidator<>(0, 100, false, false)));
 
             Setting<Integer> setting = config.get(id);
